# docs/openapi.yml
openapi: 3.0.3
info:
  title: LilyOpenCMS API
  description: |
    A comprehensive API for LilyOpenCMS content management system.
    
    ## Features
    - **Content Management**: News, albums, chapters, and media
    - **User Management**: Profiles, authentication, and permissions
    - **Public API**: Multiplatform access for mobile apps and third-party integrations
    - **SEO Management**: Comprehensive SEO tools and analytics
    - **Analytics**: Performance monitoring and user statistics
    
    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Public API endpoints (`/api/public/*`) do not require authentication.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Please implement appropriate retry logic.
    
    ## Support
    For API support, please refer to the documentation or contact the development team.
  version: 1.0.0
  contact:
    name: LilyOpenCMS Development Team
    email: support@lilyopencms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}"
    description: Custom server (set scheme/host below to try any deployment)
    variables:
      scheme:
        enum:
          - http
          - https
        default: https
      host:
        default: localhost:5000
        description: Hostname and optional port (e.g., api.example.com or localhost:5000)
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/comments/{content_type}/{content_id}:
    get:
      tags:
        - Comments
      summary: Get comments for content (news/album)
      description: Returns paginated top-level comments with reply counts
      parameters:
        - name: content_type
          in: path
          required: true
          schema:
            type: string
            enum: [news, album, chapter]
        - name: content_id
          in: path
          required: true
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200': { description: Comments returned }
        '404': { description: Not found }
  /api/comments:
    post:
      tags:
        - Comments
      summary: Create a new comment
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content_type:
                  type: string
                  enum: [news, album, album_chapter]
                content_id: { type: integer }
                content: { type: string }
                parent_id: { type: integer, nullable: true }
              required: [content_type, content_id, content]
      responses:
        '201': { description: Comment created }
  /api/comments/{comment_id}:
    put:
      tags:
        - Comments
      summary: Update comment
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
              required: [content]
      responses:
        '200': 
          description: Comment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Comment updated successfully"
                  comment:
                    $ref: '#/components/schemas/CommentInfo'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not authorized to edit this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404': 
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Comments
      summary: Delete comment
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Comment deleted successfully"
        '403':
          description: Forbidden - not authorized to delete this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404': 
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/comments/{comment_id}/like:
    post:
      tags:
        - Comments
      summary: Like/dislike comment
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [like, dislike]
              required: [action]
      responses:
        '200': { description: Reaction recorded }
        '404': { description: Not found }
  /api/comments/{comment_id}/report:
    post:
      tags:
        - Comments
      summary: Report comment
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
              required: [reason]
      responses:
        '200': { description: Report submitted }
        '404': { description: Not found }
...
components:
  schemas:
...
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
...
    CommentInfo:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserInfo'
        likes:
          type: integer
        dislikes:
          type: integer
        replies_count:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - error
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via cookies
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external ad access and third-party integrations
security:
  - sessionAuth: []
  - bearerAuth: []